+++ frontend/src/components/report/CommentReportModal/index.tsx
@@ -3,8 +3,7 @@ import React from 'react';
 import { useSelect } from '@hooks/useSelect';
 
 import Select from '@components/common/Select';
-
-import CommentModal from '../CommentModal';
+import CommentModal from '@components/post/CommentList/CommentItem/CommentModal';
 
 import { COMMENT_REPORT_MESSAGE } from './constants';
 import { CommentReportMessage } from './types';
